Move SearchTerm from cli.go to search_term.go
