+++ cli.go
@@ -64,13 +64,6 @@ type Searcher struct {
 	searchTerm        *SearchTerm
 }
 
-// SearchTerm is search term in GitHub object
-type SearchTerm struct {
-	language  string
-	filename  string
-	extension string
-}
-
 // PairList is list of Pair
 type PairList []Pair
 
@@ -139,32 +132,6 @@ func (c *CLI) Run(args []string) int {
 	searcher.output(c.outStream)
 
 	return ExitCodeOK
-}
-
-func NewSearchTerm(language string, filename string, extension string) *SearchTerm {
-	Debugf("language: %s", language)
-	Debugf("filename: %s", filename)
-	Debugf("extension: %s", extension)
-
-	return &SearchTerm{
-		language:  language,
-		filename:  filename,
-		extension: extension,
-	}
-}
-
-func (s *SearchTerm) query(keyword string) string {
-	q := keyword
-	if s.language != "" {
-		q = fmt.Sprintf("%s language:%s", q, s.language)
-	}
-	if s.filename != "" {
-		q = fmt.Sprintf("%s filename:%s", q, s.filename)
-	}
-	if s.extension != "" {
-		q = fmt.Sprintf("%s extension:%s", q, s.extension)
-	}
-	return q
 }
 
 func (s *Searcher) keywords() []string {
